+++ tests/store/store_test.py
@@ -13,26 +13,19 @@ def test_store_writer():
     )
 
     # write
-    store.write("test.txt", "test me")
+    store.save_data("test.txt", "test me")
     f = os.path.join(
-        store.config["experiment"], store.config["component"], "store", "test.txt"
+        store.config["experiment"], store.config["component"], "data", "test.txt"
     )
     assert store.filesystem.readtext(f) == "test me"
-    store.write("test.npy", np.ones([5]))
-    store.write("test.p", np.ones([5]))
-    store.write("test.json", [1, 2, 3])
-    store.write("dir/test.txt", "subdirectory")
+    store.save_data("test.npy", np.ones([5]))
+    store.save_data("test.p", np.ones([5]))
+    store.save_data("test.json", [1, 2, 3])
+    store.save_data("dir/test.txt", "subdirectory")
     f = os.path.join(
-        store.config["experiment"], store.config["component"], "store", "dir/test.txt"
+        store.config["experiment"], store.config["component"], "data", "dir/test.txt"
     )
     assert store.filesystem.readtext(f) == "subdirectory"
-    f = os.path.join(
-        store.config["experiment"], store.config["component"], "store.json"
-    )
-    store.write("test", True)
-    assert store.filesystem.readtext(f) == '{"test": true}'
-    store.write("bla", 1)
-    assert store.filesystem.readtext(f) == '{"test": true, "bla": 1}'
 
     # observations
     store.record["test"] = 1